+++ src/pbfbench/topics/plasmidness/plasclass/__init__.py
@@ -1,0 +1,1 @@
+"""PlasClass plasmidness tool."""