+++ src/pbfbench/topics/plasmidness/plasclass/core_command.sh
@@ -1,0 +1,1 @@
+python3 ${CLASSIFY_FASTA_PY} --fasta ${FASTA} --outfile ${OUTFILE} ${USER_TOOL_OPTIONS[@]}