+++ src/pbfbench/topics/plasmidness/plasgraph2/__init__.py
@@ -1,0 +1,1 @@
+"""plASgraph2 plasmidness topic tool."""